[feature julio - task: update class UserModel]

+++ Grupo4/src/main/java/com/Projeto/Integrador/Grupo4/Model/UserModel.java
@@ -17,7 +17,7 @@ public class UserModel{
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int IdClient;
+	private long IdClient;
 	@NotBlank(message = "Informe o Nome")
 	private String name;
 	@NotBlank(message = "Informe o Email")
@@ -54,11 +54,11 @@ public class UserModel{
 	}
 
 	//Metodos GET e SET
-	public int getIdClient() {
+	public long getIdClient() {
 		return IdClient;
 	}
 
-	public void setIdClient(int idClient) {
+	public void setIdClient(long idClient) {
 		IdClient = idClient;
 	}
